// src/tile-type.ts
export enum TileType {
	EMPTY = 0,
	FULL = 1,
	HALF_TOP = 2,
	HALF_RIGHT = 3,
	HALF_BOTTOM = 4,
	HALF_LEFT = 5,
	MED_PP = 6,
	MED_NP = 7,
	MED_NN = 8,
	MED_PN = 9,
	CONVEX_PP = 10,
	CONVEX_NP = 11,
	CONVEX_NN = 12,
	CONVEX_PN = 13,
	CONCAVE_PP = 14,
	CONCAVE_NP = 15,
	CONCAVE_NN = 16,
	CONCAVE_PN = 17,
	SMALL_22_PP = 18,
	SMALL_22_NP = 19,
	SMALL_22_NN = 20,
	SMALL_22_PN = 21,
	LARGE_22_PP = 22,
	LARGE_22_NP = 23,
	LARGE_22_NN = 24,
	LARGE_22_PN = 25,
	SMALL_67_PP = 26,
	SMALL_67_NP = 27,
	SMALL_67_NN = 28,
	SMALL_67_PN = 29,
	LARGE_67_PP = 30,
	LARGE_67_NP = 31,
	LARGE_67_NN = 32,
	LARGE_67_PN = 33,
	EDGE_TOP = 34,
	EDGE_RIGHT = 35,
	EDGE_BOTTOM = 36,
	EDGE_LEFT = 37,
	EDGE_CORNER_UL = 38,
	EDGE_CORNER_UR = 39,
	EDGE_CORNER_DL = 40,
	EDGE_CORNER_DR = 41,
}

export const NUM_TILE_TYPES = 42;